fix some bag with timepicker and item detail fragment

+++ app/src/main/java/com/dstakhanov/findealscommithistory/presentation/item/InstrumentItemFragment.kt
@@ -1,16 +1,19 @@
 package com.dstakhanov.findealscommithistory.presentation.item
 
 import android.content.Context
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
+import com.dstakhanov.findealscommithistory.R
 import com.dstakhanov.findealscommithistory.databinding.FragmentInstrumentItemBinding
 import com.dstakhanov.findealscommithistory.domain.item.InstrumentItem
 import com.dstakhanov.findealscommithistory.presentation.InstrumentApp
@@ -18,7 +21,7 @@ import com.dstakhanov.findealscommithistory.presentation.ViewModelFactory
 import com.dstakhanov.findealscommithistory.presentation.item.adapters.InstrumentItemListAdapter
 import javax.inject.Inject
 
-class InstrumentItemFragment : Fragment(){
+class InstrumentItemFragment : Fragment() {
     private lateinit var viewModel: InstrumentItemViewModel
     private lateinit var instrumentListAdapter: InstrumentItemListAdapter
 
@@ -59,7 +62,7 @@ class InstrumentItemFragment : Fragment(){
             instrumentListAdapter.submitList(it)
         }
         binding.buttonAddInstrumentItem.setOnClickListener {
-                launchDetailFragment(InstrumentItem.UNDEFINED_ID, MODE_ADD)
+            launchDetailFragment(InstrumentItem.UNDEFINED_ID, MODE_ADD)
         }
 
     }
@@ -68,6 +71,7 @@ class InstrumentItemFragment : Fragment(){
         super.onDestroyView()
         _binding = null
     }
+
     private fun launchDetailFragment(instrumentItemId: Int, screenMode: String) {
         findNavController().navigate(
             InstrumentItemFragmentDirections.actionInstrumentItemListToInstrumentItemDetailFragment(
@@ -97,11 +101,13 @@ class InstrumentItemFragment : Fragment(){
         setupClickListener()
         setupSwipeListener(binding.rvInstrumentItemList)
     }
+
     private fun setupClickListener() {
         instrumentListAdapter.onInstrumentItemClickListener = {
             launchDetailFragment(it.id, MODE_EDIT)
         }
     }
+
     private fun setupSwipeListener(rvInstrumentList: RecyclerView) {
         val callback = object : ItemTouchHelper.SimpleCallback(
             0,
@@ -117,7 +123,20 @@ class InstrumentItemFragment : Fragment(){
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 val item = instrumentListAdapter.currentList[viewHolder.adapterPosition]
-                viewModel.deleteInstrumentItem(item)
+                AlertDialog.Builder(requireContext()).apply {
+                    setTitle(getString(R.string.dialog_label))
+                    setMessage(getString(R.string.dialog_question))
+                    setPositiveButton(android.R.string.ok) { dialog, which ->
+                        viewModel.deleteInstrumentItem(item)
+                    }
+                    setNegativeButton(android.R.string.cancel) { dialog, which ->
+                        instrumentListAdapter.notifyItemChanged(viewHolder.adapterPosition)
+                    }
+                    setOnCancelListener { dialog ->
+                        instrumentListAdapter.notifyItemChanged(viewHolder.adapterPosition)
+                    }
+                    show()
+                }
             }
 
         }
@@ -136,7 +155,6 @@ class InstrumentItemFragment : Fragment(){
 
         private const val MODE_EDIT = "mode_edit"
         private const val MODE_ADD = "mode_add"
-        private const val MODE_UNKNOWN = ""
 
     }
